task4/back/requirements.txt
mysql-connector-python
jinja2
